+++ test/unit/pqueue/mspqueue.cpp
@@ -59,8 +59,8 @@ namespace {
             value_type * pLast = pFirst + pq.capacity();
 
             ASSERT_TRUE( pq.empty() );
-            ASSERT_EQ( pq.size(), 0 );
-            ASSERT_EQ( pq.capacity(), base_class::c_nCapacity - 1 );
+            ASSERT_EQ( pq.size(), 0u );
+            ASSERT_EQ( pq.capacity(), size_t( base_class::c_nCapacity - 1 ));
 
             size_t nSize = 0;
 
@@ -122,14 +122,14 @@ namespace {
 
             ASSERT_TRUE( !pq.full() );
             ASSERT_TRUE( !pq.empty() );
-            ASSERT_EQ( pq.size(), 1 );
+            ASSERT_EQ( pq.size(), 1u );
 
             ASSERT_TRUE( pq.pop( kv ) );
             EXPECT_EQ( kv.k, base_class::c_nMinValue );
 
             ASSERT_TRUE( !pq.full() );
             ASSERT_TRUE( pq.empty() );
-            ASSERT_EQ( pq.size(), 0 );
+            ASSERT_EQ( pq.size(), 0u );
 
             // Clear test
             for ( value_type * p = pFirst; p < pLast; ++p ) {
@@ -141,7 +141,7 @@ namespace {
             pq.clear();
             ASSERT_TRUE( pq.empty() );
             ASSERT_TRUE( !pq.full() );
-            ASSERT_EQ( pq.size(), 0 );
+            ASSERT_EQ( pq.size(), 0u );
 
             // clear_with test
             for ( value_type * p = pFirst; p < pLast; ++p ) {
@@ -156,7 +156,7 @@ namespace {
                 pq.clear_with( std::ref( disp ) );
                 ASSERT_TRUE( pq.empty() );
                 ASSERT_TRUE( !pq.full() );
-                ASSERT_EQ( pq.size(), 0 );
+                ASSERT_EQ( pq.size(), 0u );
                 ASSERT_EQ( disp.m_nCallCount, pq.capacity() );
             }
         }
